pyproject.toml
[project]
name = "robo-advisor"
version = "0.1.0"
description = "A FastAPI application with TortoiseORM backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "tortoise-orm>=0.20.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "yfinance>=0.2.37",
    "numpy>=1.26.4",
    "pandas>=2.2.1",
    "matplotlib>=3.8.3",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
